api_sync/requirements.txt
# Requirements for API Sync Package
requests>=2.28.0
google-cloud-secret-manager>=2.12.0
python-dotenv>=1.0.0
